Add files via upload
This website contains the source code for our tourism website, aiming to showcase the beauty, history, and attractions of our city to visitors and tourists. Key features include an "About Us" page with details about our mission and services, a "Contact Us" page with a form for inquiries, and a "Climate" page with essential weather information. The website also offers smooth page navigation and captivating images to entice visitors.